Third commit

+++ handlers/completTaskHandler.go
@@ -8,7 +8,7 @@ import (
 	"time"
 )
 
-// Функция CompletTaskHandler по идентификатору ставит задачу выполненной или удаляет если правило не указано
+// Функция CompletTaskHandler ставит задачу выполненной либо удаляет её
 func CompletTaskHandler(w http.ResponseWriter, r *http.Request, db *sql.DB) {
 
 	var task functions.Schedule
@@ -22,7 +22,20 @@ func CompletTaskHandler(w http.ResponseWriter, r *http.Request, db *sql.DB) {
 		return
 	}
 
-	functions.SearchTaskById(w, r, db)
+	var exists bool
+
+	err = db.QueryRow("SELECT EXISTS(SELECT 1 FROM scheduler WHERE id = ?)", task.Id).Scan(&exists)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(w).Encode(map[string]string{"error": "Ошибка при поиске задачи по идентификатору"})
+		return
+	}
+
+	if !exists {
+		w.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(w).Encode(map[string]string{"error": "Ошибка при поиске задачи"})
+		return
+	}
 
 	if task.Repeat == "" {
 		_, err = db.Exec("DELETE FROM scheduler WHERE id = ?", task.Id)
